Create rule.go
